Change n_retry to attempts
Using attempts rather than number of retries is I think less
ambigous in meaning. This is a breaking change, but as far as I
know no one is using this method right now anyway.

Because the change is (slightly) breaking and also the library
is a bit more battle-tested now (I've been using it heavily for
about 6 months in my own work), I am bumping version to
0.1.0