Add tox integration
Now if you `pip install tox`, you can just run `tox` and it will
run the test suite for python27 and python35.

Note that unlike py.test, tox prints out a warnings section if code
throws warnings, and there's a spot in the tdx test where we do so;
the warning is safe to ignore as long as all tests pass.